revert consts
