+++ src/consts/account.ts
@@ -1,6 +1,1 @@
 export const DEFAULT_ACCOUNT_LABEL = 'Account'
-
-export const STATUS_WRAPPED_METHODS = {
-  selectAccount: 'INITIAL',
-  addAccounts: 'INITIAL'
-} as const
