--- src/consts/keystore.ts
@@ -1,12 +1,0 @@
-export const STATUS_WRAPPED_METHODS = {
-  unlockWithSecret: 'INITIAL',
-  addSecret: 'INITIAL',
-  addSeed: 'INITIAL',
-  updateSeed: 'INITIAL',
-  deleteSeed: 'INITIAL',
-  removeSecret: 'INITIAL',
-  addKeys: 'INITIAL',
-  addKeysExternallyStored: 'INITIAL',
-  changeKeystorePassword: 'INITIAL',
-  updateKeyPreferences: 'INITIAL'
-} as const
