--- src/consts/storage.ts
@@ -1,3 +1,0 @@
-export const STATUS_WRAPPED_METHODS = {
-  associateAccountKeysWithLegacySavedSeedMigration: 'INITIAL'
-} as const
